use size_t for more syscalls, add -Wsign-conversion

+++ include/user/syscalls.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -9,9 +11,9 @@ int MyTid();
 int MyParentTid();
 void Exit();
 void Yield();
-int Send(int tid, const char* msg, int msglen, char* reply, int rplen);
-int Receive(int* tid, char* msg, int msglen);
-int Reply(int tid, const char* reply, int rplen);
+int Send(int tid, const char* msg, size_t msglen, char* reply, size_t rplen);
+int Receive(int* tid, char* msg, size_t msglen);
+int Reply(int tid, const char* reply, size_t rplen);
 
 #ifdef __cplusplus
 }
